Move library source files to cosmic subfolder, lots of changes.
Did a major overhaul of the Neighborhood Algorithm code.
Many other additions to the codebase. Things are mostly in a
working state, at least for the basic Be-10 and Al-26 production
code and the Neighborhood Algorithm implementation. We now have a
package installer script (setup.py) that actually works. There is
still a lot of code in the repository that is not useful and should
be removed soon.