Production after code/results review with John
Met with John Stone. After comparing results of our simulation code and
clearing up a few differences, our models produce essentially the same
results to within a few atoms after millions of years of exposure and
erosion in several different scenarios. Production code should be
considered frozen apart from superficial changes, and if it is changed
we need to be certain that it does not alter the results.