NA resampling overhaul
Extensive typing and of variables in the NA code.

Fixed a bug where the sign was wrong in the distance calculation for NA
resampling. This was recently introduced and no real simulations depended
on it. Combined na.py and na_resample.pyx to make na.pyx.

Removed numexpr as a dependency.