Massive speedup of muon code with Cython
Cython produced very fast code when applied to the tight loops in
the muon code. I fit an analytical function to the Groom 2001 Muon range
and momentum data to avoid doing linear interpolation more than is
necessary. The fit is pretty good, though I chose to make it log-log
linear right near the surface where we need a better fit. With that
correction it looks like a good fit. The fit is not so good at the
greatest depths of the data, but we don't use that part for anything
important. We assume a constant muon flux deeper than 2e5 g cm-2.