Make interpolation code work with SciPy >= 0.14
Updates to SciPy version 0.14 and later have changed
the API for UnivariateSpine to include the optional keyword
argument 'ext'. We need to save and load this in addition to
the _data tuple when we pickle the interpolation data for
ProductionSpline. These changes allow old data formats to be
loaded and these will be saved as the new data format, which is
a dict of the structure {'_data': (..., ...), 'ext': 0}.
The previous pickled data was just the _data tuple.

We also store the filename in self.filename if the interpolation
was loaded from or saved to a file.